requirements.txt
absl-py==2.0.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bootstrap4==0.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.7
Django==4.2.3
django-crispy-forms==2.0
django-environ==0.11.2
django-filter==23.2
django-import-export==3.3.6
django-taggit==4.0.0
djangorestframework==3.14.0
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
Faker==22.0.0
fastapi==0.103.2
fastjsonschema==2.18.0
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.19.4
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==2.15.0
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.5.1
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
ml-dtypes==0.2.0
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
mysql-connector-python==8.2.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.25.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pygame==2.3.0
PyGetWindow==0.0.9
Pygments==2.16.1
PyJWT==2.8.0
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-json-logger==2.0.7
pytweening==1.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.0
scipy==1.11.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
tablib==3.5.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.35.2
twilio==8.5.0
typing_extensions==4.7.1
tzdata==2023.3
unicorn==2.0.1.post1
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.23.2
virtualenv==20.24.5
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==3.0.1
widgetsnbextension==4.0.8
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
